feat: arbitrary solutions for thm/ax challenges
We remove the error "Challenge and solution constant kind don't match"
in favor of permitting any kind of solution to a challenge that is a
theorem or axiom.

Note that permitted axioms must still be declared as axioms or theorems
in the solution; it would invite inconsistency if we allowed
instantiating informative axioms with definitions.

This allows challenges that demand the construction of terms of non-Prop
types.

Fixes #3 (mostly)
Closes #5 (better alternative)